+++ examples/llm-api/quickstart_advanced.py
@@ -61,7 +61,7 @@ def add_llm_args(parser):
         default='AUTO',
         choices=[
             'AUTO', 'CUTLASS', 'TRTLLM', 'VANILLA', 'WIDEEP', 'DEEPGEMM',
-            'CUTEDSL', 'TRITON'
+            'CUTEDSL', 'TRITON', 'DENSEGEMM'
         ],
         help=
         'MoE backend to use. AUTO selects default backend based on model. It currently doesn\'t always give the best choice for all scenarios. The capabilities of auto selection will be improved in future releases.'
